% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxvalue.R
\name{boxvalue}
\alias{boxvalue}
\title{The box plot value}
\usage{
boxvalue(x)
}
\arguments{
\item{x}{vector of sample data from population 1}
}
\value{
This function returns a list of 6 components:
n: the number of observation the boxplot is drawn with

conf: upper/lower extremes of the notch, out-value of the outliers

group: a vector of the same length that
indicate to which group the outlier belongs to.

names: a vector of names for the groups.
}
\description{
The box plot value
}
\examples{
myboxplotvalue (x)
}
