% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{data set for orchard}
\format{
A data frame with 12 rows and 5 variables
\describe{
  \item{DAY}{day of the month}
  \item{CONDITION}{the time of day}
  \item{THION}{pesticide level}
  \item{OXON}{pesticide level}
  \item{RATIO}{ratio f the levels}
  ...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
ddt
}
\description{
Several measuremets for an orchard
}
\keyword{datasets}
