% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{Prediction Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{training data set}

\item{cl}{true class value of the training data}

\item{k_nn}{number of nearest neighbors}

\item{k_cv}{number of folds used in cross validation}
}
\value{
A vector of predicted class and a numeric with cross validation classification error
}
\description{
This function uitilizes cross validation and nearest validation to predict the output
}
\examples{
train = mtcars[, 2:6]
cl = mtcars[, 1]
my_knn_cv(train, cl, 5, 5)

}
