% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Function}
\usage{
my_lm(formula, dat)
}
\arguments{
\item{formula}{x and y for the model}

\item{dat}{the data set that contains x and y}
}
\value{
A table with coefficients of the model
}
\description{
This function performs linear regression on model
}
\examples{
my_lm(mpg ~ hp, mtcars)
my_lm(mpg ~ drat + wt, mtcars)

}
