% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{Student's t_test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector to be performed t.test}

\item{alternative}{Alternative hypothesis (only accepting "less", "greater", or "two.sided")}

\item{mu}{null hypothesis value of mean}
}
\value{
A list with test statistics, degree of freedom of \code{x} and the alternative
}
\description{
This function performs statistical t.test
}
\examples{
my_t.test(rnorm(10, 1, 1), "greater", 0)
my_t.test(rnorm(20), "less", 0.5)

}
