% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.r
\name{mybin}
\alias{mybin}
\title{My Binomial (Lab5)}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of runs}

\item{n}{size}

\item{p}{probability}
}
\value{
A sophisticated barplot
}
\description{
My Binomial (Lab5)
}
\examples{
\dontrun{mybin(iter=1000,n=18, p=0.3)}
}
