% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary.R
\name{mydata}
\alias{mydata}
\title{A function for descriptive statistics}
\usage{
mydata(x)
}
\arguments{
\item{x}{vector of sample data from the population}
}
\value{
The function returns mean,median,25th and 75th quartiles,min,max
}
\description{
A function for descriptive statistics
}
\examples{
mydata(x = rnorm (20,5,10))
}
