% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My Normal Curve (Lab6)}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{area}

\item{iter}{number of runs}
}
\value{
A normal graph with the given area calculater
}
\description{
My Normal Curve (Lab6)
}
\examples{
\dontrun{myncurve(mu=4,sigma=2,a=5)}
}
