% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{myplot}
\alias{myplot}
\title{A function to plot x, y}
\usage{
myplot(x, y)
}
\arguments{
\item{x}{vector of sample data from population 1}

\item{y}{vector of sample data from population 2}
}
\value{
2 plots, for sample one and two
}
\description{
A function to plot x, y
}
\examples{
x = rnorm(30,10,12); y = rnorm(40,20,15); plot(x,y)
}
