% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{My Read Function (Lab 2)}
\usage{
myread(csv)
}
\arguments{
\item{csv}{document to read}
}
\value{
Read into table
}
\description{
My Read Function (Lab 2)
}
\examples{
\dontrun{myread("ORCHARD.csv")}
}
