% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{My sample (Lab 5)}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Size}

\item{iter}{Number of iterations}

\item{time}{How long to sleep}
}
\value{
A sophisticated barplot of a sample
}
\description{
My sample (Lab 5)
}
\examples{
\dontrun{mysample(n=1000, iter=30)}
}
