% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myviolinplot.R
\name{myviolinplot}
\alias{myviolinplot}
\title{A function to create a violin plot to compare the two samples}
\usage{
myviolinplot(x, y)
}
\arguments{
\item{x}{: vector of sample data from population 1}

\item{y}{: vector of sample data from population 2}
}
\value{
A violin
}
\description{
A function to create a violin plot to compare the two samples
}
\examples{
x = rnorm(30,10,12); y = rnorm(40,20,15); myviolinplot(x,y)

}
