% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{shineyapp}
\alias{shineyapp}
\title{This is an example of a shiny package}
\usage{
shineyapp()
}
\value{
A shiny app for descriptive and inferential statistics
}
\description{
Get into reactive programming with shiny
}
\examples{
shinyapp()
}
