% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest}
\alias{ttest}
\title{A function for t tests}
\usage{
ttest(x, y, paired = FALSE, alpha = 0.05)
}
\arguments{
\item{x}{vector of sample data from population 1}

\item{y}{vector of sample data from population 2}

\item{paired}{TRUE or FALSE,
paired = FALSE by default
The samples are paired = TRUE, if each experimental unit is measured twice.
In that case the first experiments creates sample x,
and the second experiment creates sample y.
The two-sided t-test is used for paired sapmles.

The test returns an error if the samples have different size.}

\item{alpha}{alpha level is always between 0,1,
which determines the confidence interval for the P-value.}
}
\value{
The test return a list containing the data, and t test object.
}
\description{
The constructor function is based on t-test, that evaluates the NUll hypothesis H0.
H0 (The Null Hypothesis) claims that two populations have the same mean.
}
\details{
x and y have underlying normal distribution.
}
\examples{
myttest(x = rnorm(30,10,12), y = rnorm(40, 7, 10))
}
