% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinplot.R
\name{violinplot}
\alias{violinplot}
\title{A function to create a violin plot to compare the two samples}
\usage{
violinplot(x, y)
}
\arguments{
\item{x}{: vector of sample data from population 1}

\item{y}{: vector of sample data from population 2}
}
\value{
A violin
}
\description{
A function to create a violin plot to compare the two samples
}
